refactor: the benchmark for the hook now runs the hook's own extraction
`longmemeval -hook-precision` reports what the auto-recall gate would do, and it
was reproducing the gate rather than calling it. The hook turns a prompt into
terms with an identifier test, a length floor, a six-term cap and fallbacks for
CJK and Cyrillic; the benchmark used the ranking's tokenisation, which has none
of those. It also asked for eight candidates where the hook now asks for
thirty-two.

Both live in internal/prompt now, and both callers use them. The extraction is
moved, not rewritten — cmd/deja behaves exactly as before, which is why nothing
in its tests moved except the two that reached into functions that no longer
live there.

What the benchmark said before and after, on 60 cross-paired prompts whose
answer is never present, so every injection is a false fire:

  before   39/60 = 65.0%   (one term 34, two or more 20)
  after    37/60 = 61.7%   (one term 35, two or more 17)

It was overstating the gate's false-fire rate and splitting it differently
between the one-term and two-term cases. Neither number moved because the gate
changed; the gate was never what was being measured.
